+++ libft/other_functions/ft_atoi_mod.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_atoi.c                                          :+:      :+:    :+:   */
+/*   ft_atoi_mod.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: feralves < feralves@student.42sp.org.br    +#+  +:+       +#+        */
+/*   By: feralves <feralves@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/17 00:05:28 by feralves          #+#    #+#             */
-/*   Updated: 2022/10/10 19:52:36 by feralves         ###   ########.fr       */
+/*   Updated: 2023/02/15 22:33:18 by feralves         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,11 +18,11 @@
 *@return int the converted value or 0 on error
 */
 
-int	ft_atoi_mod(char *nptr)
+long int	ft_atoi_mod(char *nptr)
 {
-	int	count;
-	int	result;
-	int	signal;
+	long int	count;
+	long int	result;
+	long int	signal;
 
 	signal = 1;
 	count = 0;
